# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Gym Pulse workout tracking app.",
        "title": "Gym Pulse API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/participation": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "List participation",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.DayParticipation"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/programs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "List programs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.Program"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Create program",
                "parameters": [
                    {
                        "description": "Program",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateProgramRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Program"
                        }
                    }
                }
            }
        },
        "/api/v1/programs/from-starter": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Clone starter program",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stable operation key",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Starter selection",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CloneStarterProgramRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Program"
                        }
                    }
                }
            }
        },
        "/api/v1/programs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Get program",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Program"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Update program",
                "parameters": [
                    {
                        "description": "Program replacement",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateProgramRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Program"
                        }
                    }
                }
            }
        },
        "/api/v1/schedule": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "List schedule",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.ScheduledWorkout"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/schedule/materialize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Materialize schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stable operation key",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Date range",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MaterializeScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.ScheduledWorkout"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/schedule/regenerate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Regenerate schedule",
                "parameters": [
                    {
                        "description": "Regeneration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegenerateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RegenerateScheduleResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scheduled-workouts/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Edit scheduled workout",
                "parameters": [
                    {
                        "description": "Snapshot edit",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PatchScheduledWorkoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ScheduledWorkout"
                        }
                    }
                }
            }
        },
        "/api/v1/scheduled-workouts/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Complete scheduled workout",
                "parameters": [
                    {
                        "description": "Completion operation",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RevisionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ScheduledWorkout"
                        }
                    }
                }
            }
        },
        "/api/v1/scheduled-workouts/{id}/extra-sets": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Add extra set",
                "parameters": [
                    {
                        "description": "Extra set",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ExtraSetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ScheduledWorkout"
                        }
                    }
                }
            }
        },
        "/api/v1/scheduled-workouts/{id}/sets/{set_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "Record required set",
                "parameters": [
                    {
                        "description": "Set result",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SetMutationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ScheduledWorkout"
                        }
                    }
                }
            }
        },
...
        "/api/v1/starter-programs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goal-training"
                ],
                "summary": "List starter programs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.StarterProgram"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.DayParticipation": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "finalized_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "local_date": {
                    "type": "string"
                },
                "participated": {
                    "type": "boolean"
                },
                "revision": {
                    "type": "integer"
                },
                "scheduled_opportunity": {
                    "type": "boolean"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
        "model.Program": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "primary_goal": {
                    "type": "string"
                },
                "revision": {
                    "type": "integer"
                },
                "roadmap": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "starter_program_id": {
                    "type": "string"
                },
                "starter_version": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "workouts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProgramWorkout"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token — prefix value with \"Bearer \"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}